+++ Rdmp.Core.Tests/Curation/Integration/QueryBuildingTests/AggregateBuilderTests/MySqlAggregateBuilderTests.cs
@@ -8,6 +8,7 @@ using FAnsi;
 using NUnit.Framework;
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.QueryBuilding;
+using ReusableLibraryCode.Settings;
 
 namespace Rdmp.Core.Tests.Curation.Integration.QueryBuildingTests.AggregateBuilderTests
 {
@@ -44,11 +45,14 @@ LIMIT 32"),CollapseWhitespace(builder.SQL.Trim()));
 
             topx.DeleteInDatabase();
         }
-        [Test]
-        public void Test_AggregateBuilder_MySql_Top31OrderByCountAsc()
+        [TestCase(true)]
+        [TestCase(false)]
+        public void Test_AggregateBuilder_MySql_Top31OrderByCountAsc(bool useAliasForGroupBy)
         {
             _ti.DatabaseType = DatabaseType.MySql;
             _ti.SaveToDatabase();
+
+            UserSettings.UseAliasInsteadOfTransformInGroupByAggregateGraphs = useAliasForGroupBy;
 
             var builder = new AggregateBuilder(null, "count(*)", null);
             builder.AddColumn(_dimension1);
@@ -56,9 +60,10 @@ LIMIT 32"),CollapseWhitespace(builder.SQL.Trim()));
             var topx = new AggregateTopX(CatalogueRepository, _configuration, 31);
             topx.OrderByDirection = AggregateTopXOrderByDirection.Ascending;
             builder.AggregateTopX = topx;
-            
 
-            Assert.AreEqual(CollapseWhitespace(@"/**/
+            if (useAliasForGroupBy)
+            {
+                Assert.AreEqual(CollapseWhitespace(@"/**/
 SELECT 
 Col1,
 count(*) AS MyCount
@@ -69,9 +74,27 @@ Col1
 order by 
 MyCount asc
 LIMIT 31"), CollapseWhitespace(builder.SQL));
+            }
+            else
+            {
+                Assert.AreEqual(CollapseWhitespace(@"/**/
+SELECT 
+Col1,
+count(*) AS MyCount
+FROM 
+T1
+group by 
+Col1
+order by 
+count(*) asc
+LIMIT 31"), CollapseWhitespace(builder.SQL));
+            }
+            
 
 
             topx.DeleteInDatabase();
+
+            UserSettings.UseAliasInsteadOfTransformInGroupByAggregateGraphs = false;
         }
     }
 }